requirements.txt
python-dateutil
scikit-learn
numpy
pandas
gensim
dill
